### 2019-06-04 #55 @whedon
[PRE REVIEW]: mLEARn: An Implementation of Multi-layer Perceptron in C++
**Submitting author:** @kalu-o (<a href="http://orcid.org/0000-0002-8152-4863">Kalu U. Ogbureke</a>)
**Repository:** <a href="https://github.com/kalu-o/mLEARn" target ="_blank">https://github.com/kalu-o/mLEARn</a>
**Version:** v1.0.1
