### 2019-06-04 #55 @arfon
> I don't see anything wrong with the paper metadata ... @arfon: would you have a quick look for us?

Yeah, this is my fault. Fixing in https://github.com/openjournals/jose/pull/31
