### 2019-06-05 #55 @labarba
@dsblank — Would you be willing to contribute a review for JOSE? The submission is titled "mLEARn: An Implementation of Multi-layer Perceptron in C++" Let me know!
