### 2019-06-10 #55 @willingc
Hi @trallard, I saw your name on the list of potential reviewers and thought you would be a good reviewer. Would you be willing to contribute a review for JOSE? The submission is titled "mLEARn: An Implementation of Multi-layer Perceptron in C++" Let me know! Thanks.
