### 2019-06-24 #59 @labarba
Our [Reviewer guidelines](http://jose.theoj.org/about#reviewer_guidelines) cover briefly our expectations.
