### 2019-06-24 #55 @willingc
Thanks @labarba for starting the review.
