### 2019-08-12 #59 @labarba
@dsblank, @trallard — hello, there! Could we have a status update from you both on your review for JOSE? Thanks!
