### 2019-08-28 #59 @trallard
@kalu-o The paper has a clear statement of need, but the docs in the repository are lacking this.
I think this should be added to make this explicit in the docs.
