### 2019-08-28 #59 @dsblank
Hi all. I am having a hard time wrapping my head around the form of the publication, and how much responsibility to, say, cite previous work.
