### 2019-08-29 7d9fde3
added functionalities for psdquare optimizer
