### 2019-09-06 #59 @kalu-o
> Hi all. I am having a hard time wrapping my head around the form of the publication, and how much responsibility to, say, cite previous work.

Hi,
Did you mean citation of previous works are not sufficient? I understand from the paper guideline that technical details are not necessary.
