### 2019-09-06 #59 @kalu-o
> @kalu-o The paper has a clear statement of need, but the docs in the repository are lacking this.
> I think this should be added to make this explicit in the docs.

Hi,
Besides README.md, which other "doc" should reflect this?
