### 2019-09-12 a3bd22d
added Cifar and Eminist data set
