### 2019-09-25 #59 @labarba
> Hi all. I am having a hard time wrapping my head around the form of the publication, and how much responsibility to, say, cite previous work.

@dsblank —The publication is a summary, an "advertisement," of the scholarship contained in the software submission. The software is meant to be used in the teaching of machine learning. In the paper, the authors should cite similar efforts where others have created educational software or learning materials to teach similar topics.
