### 2019-10-23 #59 @trallard
Sorry for taking so long to complete this review.
I have gone over the functionality and performance of the package and did not find any major issues.

I opened an issue on the main repo to add a clear statement of need to the main README.
I believe that is the only outstanding item from my side.
