### 2020-02-05 fb875fb
modified layer and unit test
