### 2021-02-19 #59 @labarba
@abuccts 👋  — Hello! You signed up as a reviewer for JoSE a while back, with interest in machine learning and others. Would you be able to contribute a review for this submission?
**mLEARn: An Implementation of Multi-layer Perceptron in C++**
https://github.com/openjournals/jose-reviews/issues/59
