### 2021-04-30 #59 @labarba
hi @rasbt 👋
Thank you for agreeing to review this submission to the Journal of Open Source Education! We really appreciate it.

This submission consists of a software library for students to play around with a multi-layer perceptron and come to understand the algorithm. You have a review checklist at the top of this review thread to guide you. Have a look at the [JOSE scope](https://jose.theoj.org/about#submitting), and ask any questions you may have right here!
