### 2021-05-21 kalu-o/mLEARn#3 @rasbt
JOSE Review
Just reviewed this via https://github.com/openjournals/jose-reviews/issues/59 and found that the repo & software is in great shape! Just a few minor points:

- [ ] 1)

>  Version: Does the release version given match the GitHub release (v1.0.1)?

Based on the documentation (https://github.com/kalu-o/mLEARn/tree/master/doc), and code comments (e.g., https://github.com/kalu-o/mLEARn/blob/master/src/network.cpp), it seems to be version 1.0.0

- [ ] 2)

> This example illustrates solving a regression probplem with an MLP. The train and test datasets are "oilTrn.dat" and "oilTst.dat" located in the data/ directory. These are slightly modified version of those included with Netlab (https://www2.aston.ac.uk/eas/research/groups/ncrg/resources/netlab/how-to-use).

For example 1, could it be made more clear that there are several data files included via data.zip? Maybe

    unzip data.zip

could even be mentioned as a step in the installation / setup section.

- [ ] 3)

> Authors: Does the paper.md file include a list of authors with their affiliations?

Looks good. Just adding this to maybe double-check if the affiliations are still accurate.
