### 2021-05-21 #59 @rasbt
This submission looks really good overall, just had a few minor points that should be addressed before accepting. Posted them here: https://github.com/kalu-o/mLEARn/issues/3
