### 2021-05-23 #59 @labarba
Thanks so much for reviewing this JOSE submission, @rasbt 🙏

 @kalu-o — Thank you for your extreme patience, as JOSE nearly ground to a halt in 2020. We're just resurfacing and beginning to re-organize. Could you have a look at the minor comments? I will be glad to move this to publication finally!
