### 2021-05-23 #59 @labarba
@kalu-o — Please check the second figure in the paper: it goes outside the margin. Maybe you can rearrange the diagrams? Consider adding figure numbers and captions.
