### 2021-05-23 #59 @labarba
hi @trallard — You reviewed this JOSE submission a while back, and only left unchecked the box for Statement of Need. The paper does include this section, while the README in the repo has a brief aim, as follows:

> The aim of mLEARn is to provide a simple and extendable machine learning platform for students in courses involving C++ and machine learning.

I don't think we need to require a JOSE-style Statement of Need on the README, so let's be satisfied with the current form.

Once published, it would be nice if @kalu-o added the JOSE badge on the README!
