### 2021-06-05 #59 @labarba
@kalu-o 👋  — We have received a new review for your JOSE submission, with minor revision requests. Could you have a look and let us know when it's ready? We'll be glad to move this to publication soon.
