### 2021-06-26 5cef497
modularize layer class into DenseLayer, SigmoidLayer, etc.
