+++ include/layer.h
@@ -1,297 +1,0 @@
-/*
-MIT License
-
-Copyright (c) 2019 Kalu U. Ogbureke
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-
-Authors: Kalu U. Ogbureke
-Change Log: 01.04.2019 - Version 1.0.0
-*/
-#ifndef LAYER_H
-#define LAYER_H
-#include <iostream>
-#include <vector>
-#include "libutil.h"
-
-namespace mlearn {
-template <class T>
-/**
-    The Layer class represents a layer in a neural network.
-    The pointers "previous" and "next" point to the previous
-    and next layers respectively. Each layer has a forward
-    function(forwardProp) that produces output data from input
-    data and a backward function(backwardProp) that produces an
-    output delta (gradient) from an input delta. A layer can be
-    a hidden or an output layer.
-    \n
-    @code
-        //  Creates 2 Activation objects: hidden and output
-        Activation<double> hidden("sigmoid"), output("softmax");
-        //  A hidden layer with input and output dimensions 2
-        //  Activation function used is sigmoid
-        Layer<double> hidden_layer(2, 2, "hidden", hidden);
-        //  An output layer with input and output dimensions 2 and 1 respectively.
-        //  Activation function used is softmax
-        Layer<double> output_layer(2, 1, "output", output);
-    @endcode
-
-    @note
-    The output dimension of the last hidden layer must be same as the input dimension
-    of the output layer.
-*/
-class Layer
-{
-    protected:
-        /** Input dimension of layer */
-        uint64_t input_dim;
-        /** Output dimension of layer */
-        uint64_t output_dim;
-        /** Pointer to previous layer */
-        Layer* previous;
-        /** Pointer to next layer */
-        Layer* next;
-        /** Output data of layer */
-        NetNode<T> output_data;
-        /** Output delta of layer */
-        NetNode<T> output_delta;
-        /** Input data of layer */
-        NetNode<T> input_data;
-        /** Input delta of layer */
-        NetNode<T> input_delta;
-        /** Bias of layer */
-        NetNode<T> bias;
-        /** Delta of bias */
-        NetNode<T> delta_b;
-        /** Contains activations */
-        NetNode<T> activation;
-        /** Weight matrix of layer */
-        mublas::matrix<T> weight;
-        /** Delta of weight matrix */
-        mublas::matrix<T> delta_w;
-        /** Weight matrix used for regularization */
-        mublas::matrix<T> regularize_w;
[…]
