+++ src/layer.cpp
@@ -1,228 +1,0 @@
-/*
-MIT License
-
-Copyright (c) 2019 Kalu U. Ogbureke
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-
-Authors: Kalu U. Ogbureke
-Change Log: 21.05.2019 - Version 1.0.1
-Change Log: 01.04.2019 - Version 1.0.0
-*/
-#include "layer.h"
-
-namespace mlearn {
-template <class T>
-Layer<T>& Layer<T>::initialize()
-{
-    double lower = - sqrt(6.0/(input_dim + output_dim)) * 4;
-    double upper = - (lower);
-    using value_type = double;
-    std::random_device r;
-    std::uniform_real_distribution<value_type> distribution(
-        lower,
-        upper);
-    static std::default_random_engine generator {r()};
-    if(weight.size1() == 0) throw std::bad_alloc(); //("Matrix size cannot be zero!");
-    for (uint64_t i = 0; i < output_dim; ++i)
-    {
-        for (uint64_t j = 0; j < input_dim; ++j)
-            this->weight(i, j) = distribution(generator);
-    }
-
-    return *this;
-}
-
-template <class T>
-void Layer<T>::connect(Layer<T>& layer)
-{
-    previous = &layer;
-    layer.next = this;
-}
-
-template <class T>
-NetNode<T>& Layer<T>::getForwardInput(NetNode<T>& out)
-{
-    if (previous != nullptr) out = previous->output_data;
-    else out = input_data;
-    return out;
-}
-
-template <class T>
-NetNode<T>& Layer<T>::forwardProp()
-{
-    NetNode<T> data, out;
-    getForwardInput(data);
-    mublas::vector<T> dot_prod = mublas::prod(weight, data.getData());
-    out.setData(dot_prod);
-    activation = out + bias;
-    out = act_function.compute(activation);
-    setOutputData(out);
-    return output_data;
-}
-
-template <class T>
-NetNode<T>& Layer<T>::getBackwardInput(NetNode<T>& out)
-{
-    mublas::matrix<T> temp ;
-    if (next != nullptr)
-    {
-        temp = mublas::trans(next->weight);
-        mublas::vector<T> dot_prod = mublas::prod(temp, next->output_delta.getData());
-        out.setData(dot_prod);
-    }
-    else out = input_delta;
-    return out;
-}
-
-template <class T>
-NetNode<T>& Layer<T>::backwardProp()
-{
-    NetNode<T> delta, data;
-    getBackwardInput(delta);
-    getForwardInput(data);
[…]

[…]
