+++ src/main.cpp
@@ -130,7 +130,7 @@ int main(int argc, char** argv)
         desc.add_options()
             ("help,h", "gives information about usage")
             ("mode,m", po::value<std::string>(&mode)->default_value("train"), "program usage mode: either train or test")
-            ("optimizer,o", po::value<std::string>(&optimizer)->default_value("sgd"), "optimization algorithm used: sgd, adagrad, rmsprop")
+            ("optimizer,o", po::value<std::string>(&optimizer)->default_value("sgd"), "optimization algorithm used: sgd, adagrad, rmsprop, psdsquare")
             ("cost_function,c", po::value<std::string>(&cost_function)->default_value("crossentropy"), "cost function used: mse, mae, crossentropy")
             ("learning_rate,r", po::value<double>(&learning_rate)->default_value(0.05), "learning rate used for training")
             ("batch_size,s", po::value<uint32_t>(&batch_size)->default_value(10), "batch size used for training")
