+++ README.md
@@ -4,9 +4,7 @@ src="images/logo.png" style="background-color:rgba(0,0,0,0);" height=230 alt="mL
 </h2>
 
 # mLEARn
-An implementation of Mutli-layer Perceptron in C++. The aim of mLEARn
-is to provide a simple and extendable machine learning platform for students in courses
-involving C++ and machine learning.
+An implementation of Mutli-layer Perceptron in C++. The aim of mLEARn is to provide a simple and extendable machine learning platform for students in courses involving C++ and machine learning. There are currently available popular deep learning frameworks such as MXNet, Caffe and TensorFlow. Students often use these as off-the-shelf machine learning tools and have little or no control over the codes. One of the reasons for this is because the codes are advanced and production ready. ``mLEARn`` addresses these as it can be used as an off-the-shelf machine learning tool. Furthermore, the coding style makes it easier to apply what was learnt in machine learning/C++ courses and extend the functionalities. These make it easier to understand machine learning algorithms from the first principle and extend state-of-the-art.
 ## Table of Contents
 1. [Features](#1-features)
 2. [Building from Source](#2-building-from-source)
@@ -39,7 +37,7 @@ involving C++ and machine learning.
 * mini-batch stochastic gradient descent (SGD)
 * adagrad
 * rmsprop
-* psdsquare (will be added soon)
+* psdsquare
 
 ### miscellenous
 * L1/L2 regularization
@@ -64,15 +62,15 @@ $ cd build
 $ cmake  ..
 $ make 
 ```
-These are contents of the file "install.sh", so running the script also build the libraries and binaries.
+These are contents of the file "install.sh", so running the script also builds the libraries and binaries.
 ```bash
 $./install.sh
 ```
 ## 3. Running the mlearn Program
-The executables (mlearn and unit_test) will reside in bin/. You can run from the project directory or put the binaries in /usr/local/bin/ or execution path. Note that the programs look for the MNIST dataset in the data/ directory. So the data/ directory should be in your run directory, if you want to run unit test or train/test with MNIST. 
+The executables (mlearn and unit_test) will reside in bin/. You can run from the project directory or put the binaries in /usr/local/bin/ or execution path. Note that the programs look for the MNIST dataset in the data/ directory. So the data/ directory should be in your run/current directory, if you want to run unit test or train/test with MNIST. The MNIST dataset is included in the archive "data.zip"; there are several sample datasets included in data.zip. These dataset could be used by unzipping (unzip data.zip) in the working directory. 
 
 ### data format
-Each row of sample should be concatenation of features and labels using comma as the delimiter. The file may contain a header. Sample dataset for xor and MNIST can be found in the data/ directory. The following is a header-less xor data file, of feature and label dimensions 2 and 1 respectively. The delimiter is a comma (the last column represents the labels).
+Each row of a sample should be concatenation of features and labels using comma as the delimiter. The file may contain a header. A sample dataset for xor and MNIST can be found in the data/ directory. The following is a header-less xor data file, of feature and label dimensions 2 and 1 respectively. The delimiter is a comma (the last column represents the labels).
 ```
 1,0,1
 0,1,1
@@ -94,7 +92,7 @@ $ bin/mlearn --help
 ```
 ![](images/mlearn_help.png)
 ### example 1: regression
[…]
