### 2021-06-27 dc0007a
resized and captioned figures
