### 2021-06-27 3d6a418
resized and captioned figures
