### 2021-06-27 a1704aa
Update paper.md
Numbered and captioned figures
