+++ papers/paper.md
@@ -48,7 +48,7 @@ and exponential linear unit (ELU). The CostFunction class is responsible for obj
 ![](node.PNG)      
 *Figure 1: The Node class*
 
-![](cost_function.PNG)
+![](cost_function.PNG)  
 *Figure 2: The CostFunction class*
 
 The Network class is a classic MLP consisting of sequences of layers, i.e. one or more
@@ -57,10 +57,10 @@ reading train/test dataset into features and labels. Three different readers are
 namely MNISTReader, GenericReader and IrisReader. The Optimizer class is the base class responsible
 for training algorithms. Three optimizers are currently implemented: mini-batch stochastic gradient descent [@Kiefer:1952; @Ruder16], adaptive gradient (Adagrad) [@Duchi:2011] and root mean square propagation (RMSProp) [@Tieleman:2012].
 
-![](reader.PNG) 
+![](reader.PNG)   
 *Figure 3: The DataReader class*
 
-![](optimizer.PNG)
+![](optimizer.PNG)    
 *Figure 4: The Optimizer class*
 
 A number of new enhancements such as automatic differentiation, distributed computing and GPU support will be added in future.
