### 2021-06-27 #59 @kalu-o
> @kalu-o — Please check the second figure in the paper: it goes outside the margin. Maybe you can rearrange the diagrams? Consider adding figure numbers and captions.

@labarba The issues are now addressed. Many thanks.
