### 2021-06-27 #59 @kalu-o
> @kalu-o  — We have received a new review for your JOSE submission, with minor revision requests. Could you have a look and let us know when it's ready? We'll be glad to move this to publication soon.

Apologies for getting back late. The issues have been addressed ( I think). Many thanks to all for their time.
