### 2021-07-05 #59 @rasbt
@labarba I just looked at the feedback and file-change diffs, and it looks like all my concerns were addressed. Happy to recommend it for publication now.
