### 2021-07-05 #59 @labarba
@kalu-o — We now have two reviewers who have completed their review, and we are close to the finish line! After a minor revision in the paper (see my comments below), the final steps you will take are:

- [ ] make a tagged release on the repository, and report the version number here so I can update the metadata
- [ ] make a full deposit of the repository in Zenodo (or equivalent service), and report the archive DOI here, to enter it in the metadata
- [ ] manually edit the archive title and author list as needed to match the JOSE paper

My only comment on the paper is that I miss a passage that explains to readers how they might potentially use the software in their teaching and learning context. If you have used it within a course, it helps to describe briefly how it was used. Mostly, though, what we want is to make it easier for another instructor or a self-learner to visualize how they might reuse your artifacts. Have a look at the ["What should my paper contain?"](https://openjournals.readthedocs.io/en/jose/submitting.html#what-should-my-paper-contain) of our documentation.
