### 2021-07-12 d731915
Minor change in JOSS paper
