### 2021-07-13 037d865
minor edits
for publication in JOSE https://github.com/openjournals/jose-reviews/issues/59
