+++ papers/paper.md
@@ -28,21 +28,21 @@ be downloaded from https://github.com/kalu-o/mLEARn.
 
 # Introduction
 
-Artificial neural networks (ANNs) is an area of deep learning which has been well studied.
+Artificial neural networks (ANNs) comprise a well-studied area of deep learning.
 This is because of their importance in many areas, from autonomous driving through speech
 technologies [@Graves:2013]. The two main categories are usually recurrent (feedback)
 and feed-forward architectures [@Bishop:1995]. Deep learning is currently an active research area in machine learning.
 
 # Statement of Need
 
-There are currently available popular deep learning frameworks such as MXNet [@mxnet], Caffe [@caffe] and TensorFlow [@tensorflow]. Students often use these as off-the-shelf machine learning tools and have little or no control over the codes. One of the reasons for 
-this is because the codes are advanced and production ready. ``mLEARn`` addresses these as it can be used as an off-the-shelf machine learning tool. Furthermore, the coding style makes it easier to apply what was learnt in machine learning/C++ courses and extend the functionalities. These make it easier to understand machine learning algorithms from first principles and extend state-of-the-art.
+Currently, popular deep learning frameworks are available, such as MXNet [@mxnet], Caffe [@caffe] and TensorFlow [@tensorflow]. Students often use these as off-the-shelf machine learning tools and have little or no control over the implementation. One of the reasons for 
+this is because the codes are advanced and production ready. ``mLEARn`` addresses this and it can be used as an off-the-shelf machine learning tool. Furthermore, the coding style makes it easier to apply what was learnt in machine learning/C++ courses and extend the functionalities. This makes it easier to understand machine learning algorithms from first principles and extend the state-of-the-art.
 
 # Architecture of mLEARn
 
 The classes implemented in ``mLEARn`` are Node, NetNode, Activation, CostFunction, Layer,
 Network, DataReader and Optimizer. The Node class is the fundamental data structure
-used; and NetNode is an extension of the Node class used for multi-layer perceptron. The Activation class handles activations in the network. Currently, the functions implemented are sigmoid, tanh, rectified linear unit (ReLU), leaky ReLU, identity, softmax
+used; and NetNode is an extension of the Node class used for multi-layer perceptron. The Activation class handles activation functions in the network. Currently, the functions implemented are sigmoid, tanh, rectified linear unit (ReLU), leaky ReLU, identity, softmax
 and exponential linear unit (ELU). The CostFunction class is responsible for objective/loss functions. Cost functions implemented are mean squared error (MSE), mean absolute error (MAE) and cross entropy.
 
 ![](node.PNG)      
@@ -51,11 +51,11 @@ and exponential linear unit (ELU). The CostFunction class is responsible for obj
 ![](cost_function.PNG)  
 *Figure 2: The CostFunction class*
 
-The Network class is a classic MLP consisting of sequences of layers, i.e. one or more
+The Network class is a classic MLP consisting of sequences of layers, i.e., one or more
 hidden layers and an output layer. The DataReader class is the base class responsible for
-reading train/test dataset into features and labels. Three different readers are implemented,
+reading train/test datasets into features and labels. Three different readers are implemented,
 namely MNISTReader, GenericReader and IrisReader. The Optimizer class is the base class responsible
-for training algorithms. Three optimizers are currently implemented: mini-batch stochastic gradient descent [@Kiefer:1952; @Ruder16], adaptive gradient (Adagrad) [@Duchi:2011] and root mean square propagation (RMSProp) [@Tieleman:2012].
[…]
