### 2021-07-13 #59 @labarba
@kalu-o 👋  – please merge this PR with minor edits, and I'l proceed to publish your paper!
https://github.com/kalu-o/mLEARn/pull/4
