### 2021-07-14 #59 @kalu-o
> @kalu-o 👋 – please merge this PR with minor edits, and I'l proceed to publish your paper!
> [kalu-o/mLEARn#4](https://github.com/kalu-o/mLEARn/pull/4)

@labarba --done.
