### 2021-07-17 #59 @kalu-o
> Congratulations, @kalu-o, your JOSE paper is published! 🚀
>
> Huge and sincere thanks to the Reviewers: @trallard, @rasbt — your contribution makes this adventure possible 🙏

Thanks to the editor and reviewers for making this possible!
